feat: add

+++ omp-user-center/src/main/java/com/ahdms/user/center/bean/entity/Role.java
@@ -2,10 +2,7 @@ package com.ahdms.user.center.bean.entity;
 
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-import com.ahdms.framework.mybatis.core.Entity;
-import com.ahdms.framework.mybatis.annotation.TableBId;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 
 /**
@@ -32,5 +29,5 @@ public class Role  {
     /**
      * 
      */
-    private Integer parentId;
+    private Long parentId;
 }
